requirements.txt
requests==2.31.0
schedule==1.2.1
langchain-core==0.3.15
langchain-nvidia-ai-endpoints==0.3.4
langgraph==0.2.43
pydantic==2.9.2
python-dotenv==1.0.1
slack_sdk>=3.27.0
fastapi[standard]==0.115.6
uvicorn==0.34.0
cryptography>=41.0.0
